pyproject.toml
[tool.poetry]
name = "pivot-track"
version = "0.1.0"
description = ""
authors = ["Christoph Lobmeyer <lochr@posteo.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
shodan = "^1.31.0"
pyyaml = "^6.0.1"
common-osint-model = "^0.5.0"
typer = "^0.12.3"
pylance = "^0.11.0"

[tool.poetry.scripts]
pivottrack = "pivot_track.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
